+++ app/models/surveys/survey_encounter.rb
@@ -10,7 +10,16 @@ class SurveyEncounter < ActiveRecord::Base
   belongs_to :survey
   belongs_to :encounter
   belongs_to :user
+  belongs_to :parent_survey_encounter, class_name: "SurveyEncounter", foreign_key: 'parent_survey_encounter_id'
 
   # Model Methods
+
+  def survey_encounter_name
+    "#{self.survey.name}: #{self.encounter.name}"
+  end
+
+  def encounter_survey_name
+    "#{self.encounter.name}: #{self.survey.name}"
+  end
 
 end