+++ db/migrate/20150817193211_add_parent_survey_encounter_id_to_survey_encounters.rb
@@ -1,0 +1,6 @@
+class AddParentSurveyEncounterIdToSurveyEncounters < ActiveRecord::Migration
+  def change
+    add_column :survey_encounters, :parent_survey_encounter_id, :integer
+    add_index :survey_encounters, :parent_survey_encounter_id
+  end
+end